First Demo working. Queue Storage
First version of a queue storage demo.
Its a website that you upload an image to of handwriting.
Its posts it to queue storage. The server side dequeues it, wastes some time, asks for more, then passes it to a cognitive service to get the text out.

It does nothing with the text at the moment